backend/tests/requirements.txt
# TrustGuard Phase 2 Test Suite Dependencies

# HTTP requests
requests>=2.31.0

# AWS SDK
boto3>=1.40.0
botocore>=1.40.0

# Testing framework
pytest>=8.4.0
pytest-cov>=4.1.0

# Environment management
python-dotenv>=1.1.0

# Image generation (for mock receipts)
Pillow>=10.0.0

# JSON/data handling
pydantic>=2.12.0

# Date/time utilities
python-dateutil>=2.9.0

# CLI colors (optional, for better output)
colorama>=0.4.6
